+++ part1/base.py
@@ -1,0 +1,50 @@
+#!/usr/bin/env python3
+
+# base Controller class for common functionalities
+
+from ryu.base import app_manager
+from ryu.controller import ofp_event
+from ryu.controller.handler import MAIN_DISPATCHER, CONFIG_DISPATCHER, set_ev_cls
+from ryu.ofproto import ofproto_v1_3
+from ryu.lib.packet import packet
+from ryu.lib.packet import ethernet
+from ryu.lib.packet import ether_types
+
+class BaseController(app_manager.RyuApp):
+    """Base Controller skeleton for Hub and Learning Switch"""
+    OFP_VERSIONS = [ofproto_v1_3.OFP_VERSION] # openflow version
+
+    def __init__(self, *args, **kwargs):
+        super(BaseController, self).__init__(*args, **kwargs)
+        # mac_to_port[switch_dpid][mac] = port
+        self.mac_to_port = {}
+        print("Hi. Initializing Base Controller class.")
+
+    @set_ev_cls(ofp_event.EventOFPSwitchFeatures, CONFIG_DISPATCHER)
+    def switch_features_handler(self, ev):
+        """Install a table-miss flow entry so unmatched packets are sent to controller"""
+        datapath = ev.msg.datapath
+        ofproto = datapath.ofproto
+        parser = datapath.ofproto_parser
+
+        # Match everything (table-miss)
+        match = parser.OFPMatch()
+        actions = [parser.OFPActionOutput(ofproto.OFPP_CONTROLLER,
+                                          ofproto.OFPCML_NO_BUFFER)]
+        self.add_flow(datapath, 0, match, actions)
+
+    def add_flow(self, datapath, priority, match, actions, buffer_id=None):
+        """Helper to install a flow on a switch"""
+        ofproto = datapath.ofproto
+        parser = datapath.ofproto_parser
+
+        inst = [parser.OFPInstructionActions(ofproto.OFPIT_APPLY_ACTIONS, actions)]
+
+        if buffer_id:
+            mod = parser.OFPFlowMod(datapath=datapath, buffer_id=buffer_id,
+                                    priority=priority, match=match,
+                                    instructions=inst)
+        else:
+            mod = parser.OFPFlowMod(datapath=datapath, priority=priority,
+                                    match=match, instructions=inst)
+        datapath.send_msg(mod)
