--- part1/controllers.py
@@ -1,185 +1,0 @@
-# from ryu.base import app_manager
-# from ryu.controller import ofp_event
-# from ryu.controller.handler import MAIN_DISPATCHER, CONFIG_DISPATCHER, set_ev_cls
-# from ryu.ofproto import ofproto_v1_3
-
-
-# class BaseController(app_manager.RyuApp):
-#     """Base Controller skeleton for Hub and Learning Switch"""
-#     OFP_VERSIONS = [ofproto_v1_3.OFP_VERSION]
-
-#     def __init__(self, *args, **kwargs):
-#         super(BaseController, self).__init__(*args, **kwargs)
-#         self.mac_to_port = {}
-
-#     @set_ev_cls(ofp_event.EventOFPSwitchFeatures, CONFIG_DISPATCHER)
-#     def switch_features_handler(self, ev):
-#         """Handle switch features (install table-miss entry later)"""
-#         pass
-
-#     @set_ev_cls(ofp_event.EventOFPPacketIn, MAIN_DISPATCHER)
-#     def _packet_in_handler(self, ev):
-#         """Handle packets sent to controller by switch"""
-#         pass
-
-
-# class HubController(BaseController):
-#     """Hub Controller skeleton"""
-#     def __init__(self, *args, **kwargs):
-#         super(HubController, self).__init__(*args, **kwargs)
-
-#     @set_ev_cls(ofp_event.EventOFPPacketIn, MAIN_DISPATCHER)
-#     def _packet_in_handler(self, ev):
-#         # TODO: Implement flooding / controller-based forwarding
-#         pass
-
-
-# class LearningSwitch(BaseController):
-#     """Learning Switch skeleton"""
-#     def __init__(self, *args, **kwargs):
-#         super(LearningSwitch, self).__init__(*args, **kwargs)
-
-#     @set_ev_cls(ofp_event.EventOFPPacketIn, MAIN_DISPATCHER)
-#     def _packet_in_handler(self, ev):
-#         # TODO: Implement MAC learning and flow installation
-#         pass
-
-
-#!/usr/bin/env python3
-"""
-Ryu Controllers for Part 1:
- - Hub Controller
- - Learning Switch
-"""
-
-from ryu.base import app_manager
-from ryu.controller import ofp_event
-from ryu.controller.handler import MAIN_DISPATCHER, CONFIG_DISPATCHER, set_ev_cls
-from ryu.ofproto import ofproto_v1_3
-from ryu.lib.packet import packet
-from ryu.lib.packet import ethernet
-from ryu.lib.packet import ether_types
-
-
-class BaseController(app_manager.RyuApp):
-    """Base Controller skeleton for Hub and Learning Switch"""
-    OFP_VERSIONS = [ofproto_v1_3.OFP_VERSION]
-
-    def __init__(self, *args, **kwargs):
-        super(BaseController, self).__init__(*args, **kwargs)
-        # mac_to_port[switch_dpid][mac] = port
-        self.mac_to_port = {}
-
-    @set_ev_cls(ofp_event.EventOFPSwitchFeatures, CONFIG_DISPATCHER)
-    def switch_features_handler(self, ev):
-        """Install a table-miss flow entry so unmatched packets are sent to controller"""
-        datapath = ev.msg.datapath
-        ofproto = datapath.ofproto
-        parser = datapath.ofproto_parser
-
-        # Match everything (table-miss)
-        match = parser.OFPMatch()
-        actions = [parser.OFPActionOutput(ofproto.OFPP_CONTROLLER,
-                                          ofproto.OFPCML_NO_BUFFER)]
-        self.add_flow(datapath, 0, match, actions)
-
-    def add_flow(self, datapath, priority, match, actions, buffer_id=None):
-        """Helper to install a flow on a switch"""
-        ofproto = datapath.ofproto
-        parser = datapath.ofproto_parser
-
-        inst = [parser.OFPInstructionActions(ofproto.OFPIT_APPLY_ACTIONS, actions)]
-
-        if buffer_id:
-            mod = parser.OFPFlowMod(datapath=datapath, buffer_id=buffer_id,
-                                    priority=priority, match=match,
-                                    instructions=inst)
-        else:
-            mod = parser.OFPFlowMod(datapath=datapath, priority=priority,
-                                    match=match, instructions=inst)
-        datapath.send_msg(mod)
-
-
-class HubController(BaseController):
-    """Hub Controller: Floods all packets (acts like a hub)"""
-
-    @set_ev_cls(ofp_event.EventOFPPacketIn, MAIN_DISPATCHER)
-    def _packet_in_handler(self, ev):
-        msg = ev.msg
-        datapath = msg.datapath
-        ofproto = datapath.ofproto
-        parser = datapath.ofproto_parser
-        in_port = msg.match['in_port']
-
-        pkt = packet.Packet(msg.data)
-        eth = pkt.get_protocol(ethernet.ethernet)
-        if eth.ethertype == ether_types.ETH_TYPE_LLDP:
-            # Ignore LLDP packets (used by topology discovery)
-            return
-
-        dst = eth.dst
-        src = eth.src
-        dpid = datapath.id
-
-        self.logger.info("Hub: Packet in switch %s src=%s dst=%s in_port=%s",
-                         dpid, src, dst, in_port)
-
-        # Flood: send to all ports except input port
-        actions = [parser.OFPActionOutput(ofproto.OFPP_FLOOD)]
-
-        # Send packet out
-        out = parser.OFPPacketOut(datapath=datapath,
-                                  buffer_id=msg.buffer_id,
-                                  in_port=in_port,
-                                  actions=actions,
-                                  data=msg.data if msg.buffer_id == ofproto.OFP_NO_BUFFER else None)
-        datapath.send_msg(out)
-
-
-class LearningSwitch(BaseController):
-    """Learning Switch: Learns MAC-port mapping and installs flows"""
-
-    @set_ev_cls(ofp_event.EventOFPPacketIn, MAIN_DISPATCHER)
-    def _packet_in_handler(self, ev):
-        msg = ev.msg
-        datapath = msg.datapath
-        ofproto = datapath.ofproto
-        parser = datapath.ofproto_parser
-        in_port = msg.match['in_port']
-
-        pkt = packet.Packet(msg.data)
-        eth = pkt.get_protocol(ethernet.ethernet)
-        if eth.ethertype == ether_types.ETH_TYPE_LLDP:
-            return
-
-        dst = eth.dst
-        src = eth.src
-        dpid = datapath.id
-        self.mac_to_port.setdefault(dpid, {})
-
-        # Learn the source MAC → port mapping
-        self.mac_to_port[dpid][src] = in_port
-
-        if dst in self.mac_to_port[dpid]:
-            out_port = self.mac_to_port[dpid][dst]
-        else:
-            out_port = ofproto.OFPP_FLOOD
-
-        actions = [parser.OFPActionOutput(out_port)]
-
-        # Install flow rule if destination is known
-        if out_port != ofproto.OFPP_FLOOD:
-            match = parser.OFPMatch(in_port=in_port, eth_dst=dst, eth_src=src)
-            if msg.buffer_id != ofproto.OFP_NO_BUFFER:
-                self.add_flow(datapath, 1, match, actions, msg.buffer_id)
-                return
-            else:
-                self.add_flow(datapath, 1, match, actions)
-
-        # Send packet out (controller forwards the packet once)
-        out = parser.OFPPacketOut(datapath=datapath,
-                                  buffer_id=msg.buffer_id,
-                                  in_port=in_port,
-                                  actions=actions,
-                                  data=msg.data if msg.buffer_id == ofproto.OFP_NO_BUFFER else None)
-        datapath.send_msg(out)
